requirements.txt
google-cloud-aiplatform
google-genai
google-cloud-storage
python-dotenv
psycopg2-binary